The Limit
This is as far as I got before the deadline. Most things seem to work
fine. The only problem I found among the implemented features was that
while loops ended up behaving just like if statements. I spent a while
on that but could not figure it out. Also, I regret not giving myself
time to give it a visual make over. Perhaps I will update that next
semester to match the theme I choose for the OS.